Refactoring

+++ src/core/Rendering/Vulkan/VulkanMesh.h
@@ -148,9 +148,9 @@ struct TransformData
 
 struct Drawable
 {
-	uint32_t id = 0;
-	uint32_t mesh_id = 0;
-	uint32_t material_id = 0;
+	size_t id = 0;
+	size_t mesh_id = 0;
+	size_t material_id = 0;
 	bool visible = true;	// True if it is rendered by the forward/deferred renderer.
 	bool has_primitives() const;
 	const VulkanMesh& get_mesh() const;
